+++ drivers/misc/mediatek/iommu/iommu_secure.h
@@ -6,6 +6,6 @@
 #define IOMMU_SECURE_H
 
 int mtk_iommu_sec_bk_init(uint32_t type, uint32_t id);
-int mtk_iommu_sec_bk_tf(uint32_t type, uint32_t id, u32 *iova, u32 *pa, u32 *fault_id);
+int mtk_iommu_sec_bk_tf(uint32_t type, uint32_t id, u64 *iova, u64 *pa, u32 *fault_id);
 
 #endif
